updating todos rest api to make use of the user

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import * as yup from "yup";
@@ -37,9 +38,12 @@ const postSchema = yup.object({
   complete: yup.boolean().optional().default(false),
 });
 export async function POST(request: Request) {
+  const user = await getUserServerSession();
+  if (!user) return NextResponse.json({ message: "unauthorized" }, { status: 401 });
+
   try {
     const { complete, description } = await postSchema.validate(await request.json());
-    const todo = await prisma.todo.create({ data: { complete: complete, description: description } });
+    const todo = await prisma.todo.create({ data: { complete: complete, description: description, userId: user?.id } });
 
     return NextResponse.json({ todo });
   } catch (error) {
@@ -48,8 +52,11 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserServerSession();
+  if (!user) return NextResponse.json({ message: "unauthorized" }, { status: 401 });
+
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } });
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
 
     return NextResponse.json({ message: "[success]" }, { status: 200 });
   } catch (error) {
